test: add source-owned network integrity hook
