+++ src/ui/authorityIntegrity.ts
@@ -1,0 +1,20 @@
+import { invoke } from "@tauri-apps/api/core";
+
+export type AuthorityIntegrityProbeResult = {
+  adapter_id: string;
+  endpoint: string;
+  dependency_path_reached: boolean;
+};
+
+/**
+ * Source-owned integrity hook for an isolated test build. Production UI does
+ * not import this module, and the matching Tauri command exists only when the
+ * `authority-integrity-test-hooks` Cargo feature is enabled.
+ */
+export function probeAuthorityIntegrityAdapter(
+  endpoint: string,
+): Promise<AuthorityIntegrityProbeResult> {
+  return invoke<AuthorityIntegrityProbeResult>("authority_integrity_probe_adapter", {
+    input: { endpoint },
+  });
+}
